+++ packages/view-types/scatterplot-embedding/src/EmbeddingScatterplotSubscriber.js
@@ -26,6 +26,7 @@ import {
   getPointSizeDevicePixels,
   getPointOpacity,
 } from '@vitessce/scatterplot';
+import { Legend } from '@vitessce/legend';
 import { ViewType, COMPONENT_COORDINATION_TYPES } from '@vitessce/constants-internal';
 
 /**
@@ -347,6 +348,16 @@ export function EmbeddingScatterplotSubscriber(props) {
         getObsInfo={getObsInfo}
       />
       )}
+      <Legend
+        visible
+        theme={theme}
+        featureType={featureType}
+        featureValueType={featureValueType}
+        obsColorEncoding={cellColorEncoding}
+        featureSelection={geneSelection}
+        featureValueColormap={geneExpressionColormap}
+        featureValueColormapRange={geneExpressionColormapRange}
+      />
     </TitleInfo>
   );
 }